Database Updated

+++ app.js
@@ -15,8 +15,8 @@ const createTableIfNotExists = async () => {
             id INT AUTO_INCREMENT PRIMARY KEY,
             name VARCHAR(255) NOT NULL,
             address VARCHAR(255),
-            latitude DOUBLE,
-            longitude DOUBLE
+            latitude FLOAT,
+            longitude FLOAT
         );
     `;
